Added json endpoints and readme

+++ project.py
@@ -1,7 +1,8 @@
-from flask import Flask, render_template, request, redirect, jsonify, url_for, flash
+from flask import Flask, render_template, request, redirect, jsonify, url_for
+from flask import flash
 from sqlalchemy import create_engine, asc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Categories, Items,User
+from database_setup import Base, Categories, Items, User
 from flask import session as login_session
 import random
 import string
@@ -41,9 +42,6 @@ session = DBSession()
 #         return user.id
 #     except:
 #         return None
-
-
-
 
 @app.route('/login/')
 def login_s():
@@ -201,6 +199,8 @@ def gdisconnect():
             json.dumps('Failed to revoke token for given user.', 400))
         response.headers['Content-Type'] = 'application/json'
         return response
+
+
 @app.route('/logout/')
 def logout():
     if 'username' in login_session:
@@ -215,16 +215,30 @@ def logout():
     else:
         return redirect(url_for("login_s"))
 
+@app.route('/catalog/JSON')
+def catjson():
+    cats = session.query(Categories).all()
+    return jsonify(Category= [i.serialize for i in cats])
+
+
+@app.route('/catalog/<cat_name>/JSON')
+def itemsjson(cat_name):
+    cat = session.query(Categories).filter_by(name=cat_name).one()
+    items = session.query(Items).filter_by(cat_id=cat.id).all()
+    return jsonify(Items= [i.serialize for i in items])
+
 
 @app.route('/')
 @app.route('/catalog/')
 def Display_categories():
     categories = session.query(Categories).order_by(asc(Categories.name))
     # if 'username' not in login_session:
-    #     return render_template('publiccategories.html',categories=categories)
+    # return render_template('publiccategories.html', categories=categories)
     # else:
-    return render_template('categories.html',categories=categories)
-@app.route('/catalog/new/',methods=['GET','POST'])
+    return render_template('categories.html', categories=categories)
+
+
+@app.route('/catalog/new/', methods=['GET', 'POST'])
 def Add_category():
     if 'username' not in login_session:
         return redirect('/login')
@@ -237,54 +251,87 @@ def Add_category():
     else:
         return render_template('newcategory.html')
 
+
 @app.route('/catalog/<cat_name>/')
 @app.route('/catalog/<cat_name>/items/')
 def Display_Items(cat_name):
     category = session.query(Categories).filter_by(name=cat_name).one()
     items = session.query(Items).filter_by(cat_id=category.id).all()
-    return render_template('items.html',category=category,items=items)
-@app.route('/catalog/<cat_name>/items/new/',methods=['GET','POST'])
+    return render_template('items.html', category=category, items=items)
+
+
+@app.route('/catalog/<cat_name>/items/new/', methods=['GET', 'POST'])
 def Add_item(cat_name):
     if 'username' not in login_session:
         return redirect('/login')
     categories = session.query(Categories).order_by(asc(Categories.name))
     cat = session.query(Categories).filter_by(name=cat_name).one()
+    # if login_session['user_id'] != cat.user_id:
+    #  return "<script>function myFunction() {alert('You are not authorized'+
+    # ' to delete menu items to this restaurant. Please create your own'+
+    # ' restaurantin order to delete items.');}</script>"
+
     if request.method == 'POST':
-        newitem = Items(name=request.form['name'], description=request.form['description'],cat_id=cat.id)
+        newitem = Items(
+            name=request.form['name'],
+            description=request.form['description'], cat_id=cat.id
+            )
         session.add(newitem)
         session.commit()
-        return redirect(url_for('Display_Items',cat_name=cat_name))
+        return redirect(url_for('Display_Items', cat_name=cat_name))
     else:
-        return render_template('additem.html',cat=cat)
+        return render_template('additem.html', cat=cat)
+
+
 @app.route('/catalog/<cat_name>/items/<item_name>/')
-def Display_Item(cat_name,item_name):
-    category=session.query(Categories).filter_by(name=cat_name).one()
+def Display_Item(cat_name, item_name):
+    category = session.query(Categories).filter_by(name=cat_name).one()
     items = session.query()
-    return render_template('showitem.html',item=items,category=category)
-@app.route('/catalog/<cat_name>/items/<item_name>/edit/',methods=['GET','POST'])
-def Edit_item(cat_name,item_name):
+    return render_template('showitem.html', item=items, category=category)
+
+
+@app.route(
+    '/catalog/<cat_name>/items/<item_name>/edit/',
+    methods=['GET', 'POST']
+    )
+def Edit_item(cat_name, item_name):
     if 'username' not in login_session:
         return redirect('/login')
-    item = session.query(Items).filter_by(name=item_name).one()
+    item = session.query(Items).filter_by(name=item_name).all()
     categories = session.query(Categories).order_by(asc(Categories.name))
     category = session.query(Categories).filter_by(name=cat_name).one()
+    if login_session['user_id'] != item.user_id:
+        return "<script>function myFunction() +{alert('You are not authorized"
+        " to delete menu items to this restaurant. Please create your own"
+        " restaurant in order to delete items.');}"
+        "</script><body onload='myFunction()''>"
+
     if request.method == 'POST':
-        item.name=request.form['name']
-        item.description = request.form['descripton']
+        item.name = request.form['name']
+        item.description = request.form['description']
         session.add(item)
         session.commit()
         flash('Item Successfully Edited')
-        return redirect(url_for('Display_Item', cat_name=cat_name,item_name=request.form['name']))
+        return redirect(url_for('Display_Item', cat_name=cat_name,
+                        item_name=request.form['name']))
     else:
-        return render_template('edititem.html',category=category, item=item)
-@app.route('/catalog/<cat_name>/items/<item_name>/delete/',methods=['GET','POST'])
-def Delete_item(cat_name,item_name):
+        return render_template('edititem.html', category=category, item=item)
+
+
+@app.route(
+    '/catalog/<cat_name>/items/<item_name>/delete/',
+    methods=['GET', 'POST']
+    )
+def Delete_item(cat_name, item_name):
     if 'username' not in login_session:
         return redirect('/login')
     category = session.query(Categories).filter_by(name=cat_name).one()
     item = session.query(Items).filter_by(name=item_name).one()
-    # if login_session['user_id'] != restaurant.user_id:
-    #     return "<script>function myFunction() {alert('You are not authorized to delete menu items to this restaurant. Please create your own restaurant in order to delete items.');}</script><body onload='myFunction()''>"
+    if login_session['user_id'] != item.user_id:
+        return "<script>function myFunction() +{alert('You are not authorized"
+        " to delete menu items to this restaurant. Please create your own"
+        " restaurant in order to delete items.');}"
+        "</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(item)
         session.commit()
